==> codes/6_spatial_dependence_models.ipynb <==
# 7. Spatial dependence models

**Warning! This chapter is realized in R**

In this chapter we create spatial dependence econometric models to answer our research hypothesis. We believe based on spatial autocorrelation tests (performed in previous chapters) that spatial autocorrelation exists in the analyzed phenomenon. We create models for Warsaw and Cracow separately. We utilize here 1km2 grid!

Generally, we specific following functional form of the models:
$inpost_i = \alpha * \text{logistic_companies}_i + \beta * \text{demographic_variables}_i + \gamma * \text{point_of_interest_variables}_i + \theta * \text{spatial_terms}_i + \epsilon$, 

where:

* logistic_companies cover: DHL, DPD, Fedex, Poczta Polska, Ruch and UPS
* demographic_variables cover: total density, density age 0-14, density age 15-64, density age 65-...,  total mail density, total female density, female density age 0-14, female density age 15-64, female density age 65-..., male density age 0-14, male density age 15-64, male density age 65-..., feminization ratio
* point_of_interest_variables cover: buildings, shops, parks, forests, schools, railways, cycleways, parkings, crossings, bus stops

Obviously, a significant proportion of these variables will be removed from the final model due to high co-linearity or statistical insignificance, however, based on the literature and our intuition, we assume that at least 1-2 variables from each category will ultimately be significant. 

Our expectations for the results are as follows (taking into account the research hypotheses). Based on the visual analysis of spatial data carried out in the previous chapters of this work, it seems to us that InPost has deployed parcel pick-up points in accordance with the competition. Therefore, we expect the variables in the logistic_companies category to be significant and have a positive sign. In addition, it seems to us that the control variables (groups: demographic_variables, point_of_interest_variables) should also be important in the models and their sign should follow the logic: the more shops in the grid, the more parcel machines, the greater the afforestation, the fewer parcel machines, the more bus stops the more parcel machines, the higher the population density, the more parcel machines etc. In addition, we believe that extending the OLS model with spatial components is a good idea and models that take into account spillover effects will be econometrically better. We apply these expectations to both Warsaw and Cracow.

We adopted the following modeling strategy due to the complexity of the problem:

1. OLS model estimation to establish an initial functional form that is free of collinearity (VIF statistic) of the variables along with model diagnostics.
2. Estimation of the final OLS model using Stepwise Regression procedure (both directions).
3. Estimation of spatial models with 3, 2, 1 spatial components (using obtimate functional form from point 2).
4. Comparing the statistical properties of the models from step 3 and selecting the target model. We look for model with significant spatial components, highest number of significant variables and minimum AIC (we also take into account BIC).
5. For the obtimate model in terms of spatial components, we test other combinations of variables using the general to specific (LSE) procedure. 
6. Interpret the results of the model from point 5 


We covered all possible options to improve the models:
-	Changing spatial components (done - all possible components tried)
-	Changing the spatial weights matrix (done - QUEEN and ROOK tested)
-	Changing variables in the model (done - general to specific approach to make a choice for OLS and then infer based on it)

We set 10% significance level!

## Import dependencies

library(rgdal)
library(spdep)
library(tidyverse)
library(car)
library(lmtest)
library(arules)
library(spatialreg)
library(texreg)
library(stargazer)

## Models for Warsaw

### Data loading, spatial transformations and additional feature engineering

df <- readOGR("../datasets/preprocessed_data/df_warszawa.shp")
df_data <- read.csv("../datasets/preprocessed_data/df_warszawa.csv")
df <- spTransform(df, CRS("+proj=longlat +ellps=GRS80 +no_defs"))
df.cont.nb <- poly2nb(as(df, "SpatialPolygons"), queen=TRUE) #QUEEN but ROOK was tried!
df.cont.listw <- nb2listw(df.cont.nb, style="W")

cat(colnames(df_data))

We investigate again correlation between logistic companies. We can clearly see that Poczta Polska is highly collinear with other variables so we will remove it from the further analysis. The rest variables seems to be ok.

stats::cor(df_data %>% select('dhl', 'dpd', 'fedex', 'inpost', 'poczta', 'ruch', 'ups'))

Potential enhancement of the features.

df$tot_0_14 <- df$tot_0_14/df$tot
df$tot_15_64 <- df$tot_15_64/df$tot
df$tot_65__ <- df$tot_65__/df$tot

df$male_0_14 <- df$male_0_14/df$tot_male 
df$male_15_64 <- df$male_15_64/df$tot_male 
df$male_65__ <- df$male_65__/df$tot_male 

df$fem_0_14 <- df$fem_0_14/df$tot_fem 
df$fem_15_64 <- df$fem_15_64/df$tot_fem 
df$fem_65__ <- df$fem_65__/df$tot_fem 

### OLS models

Based on many tries we decided that below formula is the most appropriate general formula for the given problem. We removed Poczta Polska from logistic companies, we left total density and feminization ratio in case of demographic data and finally we leave all data regarding points of interest.

formula = inpost ~ dhl + dpd + fedex + ruch + ups +
            log(tot + 1) + fem_ratio + 
            buildings + shops + parks + forests + schools + 
            railways + cycleways + parkings + crossings + bus_stops

We estimate OLS model and we perform its diagnostics. We see that nearly 50% of the variables are significant (2 or more per each category of data). We do not have now high collinearity. However this model did not pass Ramsey and Breusch–Pagan test so the functional form is wrong and we have got heteroscedasticity in the errors. It occurred that at the level of 10% significance our residuals are spatially dependent!

ols <- lm(formula, data=df)
summary(ols)
car::vif(ols)
bptest(ols) 
resettest(ols, power=2, type="regressor")
lm.morantest(ols, df.cont.listw) 

Then we applied Stepwise Regression procedure to obtain the most significant model starting from the function form defined above.

intercept_only <- lm(inpost~1, data=df)
all <- lm(formula, data=df)
both <- step(intercept_only, direction='both', scope=formula(all), trace=0)
both

Based on Stepwise Regression we obtained final formula of the model and we are able to run our final OLS with diagnostic part. We see that nearly all variables in the model are significant (of course due to the heteroscedasticity we use biased error variance estimator - we omitted application of the robust matrix due to lack of time). There is no collinearity in the model, but again we did not pass Ramsey and B-P tests. What is more residuals from the model are still spatially dependent. In case of  interpretation we can see that: logistic companies have positive sign as expected; the greater the total density the more InPost pick-up points are stated; fem_ratio results is interesting because its sign is negative (more woman less pick-up points, quite spurious?!), in addition, the negative relationship between parks and pickup points is interesting but maybe intuitive.

formula = inpost ~ dhl + schools + log(tot + 1) + ruch + parks + 
    ups + crossings + fem_ratio + fedex
ols <- lm(formula, data=df)
summary(ols)
cat("AIC", AIC(ols))
car::vif(ols)
bptest(ols) 
resettest(ols, power=2, type="regressor")
lm.morantest(ols, df.cont.listw) 

### Spatial models with 3, 2, or 1 spatial components

Now we moved to the estimation of the spatial models. In the first step we estimate all possible models due to spatial terms and then we analyze which model is reasonable (based on number of significant variables, significant spatial terms and AIC+BIC).

GNS <- sacsarlm(formula, data=df, listw=df.cont.listw, type="sacmixed", method="LU")
summary(GNS)

SAC <- sacsarlm(formula, data=df, listw=df.cont.listw)
summary(SAC)

SDM <- errorsarlm(formula, data=df, listw=df.cont.listw, etype="mixed")
summary(SDM)

SDEM <- errorsarlm(formula, data=df, listw=df.cont.listw, etype="emixed")
summary(SDEM)

SEM <- errorsarlm(formula, data=df, listw=df.cont.listw)
summary(SEM)

SAR <- lagsarlm(formula, data=df, listw=df.cont.listw)
summary(SAR)

SLX <- lmSLX(formula, data=df, listw=df.cont.listw)
summary(SLX)

Based on the above outputs we can clearly state that the best spatial model is **SAR (Spatial Lag Model)** - significant spatial terms and many significant variables.

Then we compare models BIC and AIC. We can clearly see that the best model in comparison to the OLS is SAR, however its BIC is greater that OLS BIC. In case of AIC which is more important the best model is SLX. However based on its output we can claim that SAR is still better in case of econometric inference.

cat("ols",BIC(ols) ,'\n')
cat("GNS",BIC(GNS),'\n')
cat("SAC",BIC(SAC),'\n')
cat("SDM",BIC(SDM),'\n')
cat("SDEM",BIC(SDEM),'\n')
cat("SAR",BIC(SAR),'\n')
cat("SLX",BIC(SLX),'\n')
cat("SEM",BIC(SEM),'\n')
cat('\n')
cat("ols",AIC(ols) ,'\n')
cat("GNS",AIC(GNS),'\n')
cat("SAC",AIC(SAC),'\n')
cat("SDM",AIC(SDM),'\n')
cat("SDEM",AIC(SDEM),'\n')
cat("SAR",AIC(SAR),'\n')
cat("SLX",AIC(SLX),'\n')
cat("SEM",AIC(SEM),'\n')

Now we wanted to find better functional form for the SAR model in the general to specific procedure (LSA) done by hand, however we were not able to improve this model (our previous functional form was the best one). Residuals in the model are not autocorrelated.

formula_max = inpost ~ dhl + dpd + fedex + ruch + ups +
            log(tot + 1) + fem_ratio + 
            buildings + shops + parks + forests + schools + 
            railways + cycleways + parkings + crossings + bus_stops

formula_gen_to_specifc = inpost ~ dhl + fedex + ruch + ups +
            log(tot + 1) + fem_ratio + 
            parks + schools + 
            crossings

SAR_g2s<- lagsarlm(formula_gen_to_specifc, data=df, listw=df.cont.listw)
summary(SAR_g2s) #the same as previously!

### Direct and indirect impacts

Now we calculate direct and indirect impacts of the model. And we finally analyze its output.

Let's start with logistic companies variables. We see that either in case of direct and indirect impacts all parameters for logistic companies (apart indirect Fedex) are significant (at least around 10%). Internalization effects are greater than spill-over, but still spill-over effects are significant. In case of demographic variables we claim that total population density is important variable for the model in case of direct and indirect impact (has positive sign). Feminization ratio in case of direct impact is also significant and has negative sign (it is quite non intuitive). In case of point of interest data we see that number of schools, number of parks and number of crossings are significant in the model (crossings spillover effect is not relevant) with respectively positive, negative and positive sign. We claim that it might be intuitive.

# distribution of total impact 
# vector of traces of powers of a spatial weights matrix
# converting censored listw to CsparseMatrix form
W.c <- as(as_dgRMatrix_listw(df.cont.listw), "CsparseMatrix") 
# the default values for the number of powers is 30
trMat<-trW(W.c, type="mult") 

SAR_imp<-impacts(SAR, tr=trMat, R=200)
summary(SAR_imp, zstats=TRUE, short=TRUE)

## Models for Cracow

For the Cracow we applied the same procedure as for Warsaw, so we will omit here additional comments and focus only on the results.

### Data loading, spatial transformations and additional feature engineering

df <- readOGR("../datasets/preprocessed_data/df_krakow.shp")
df_data <- read.csv("../datasets/preprocessed_data/df_krakow.csv")
df <- spTransform(df, CRS("+proj=longlat +ellps=GRS80 +no_defs"))
df.cont.nb <- poly2nb(as(df, "SpatialPolygons"), queen=TRUE) #QUEEN but ROOK was tried!
df.cont.listw <- nb2listw(df.cont.nb, style="W")

cat(colnames(df_data))

stats::cor(df_data %>% select('dhl', 'dpd', 'fedex', 'inpost', 'poczta', 'ruch', 'ups'))

df$tot_0_14 <- df$tot_0_14/df$tot
df$tot_15_64 <- df$tot_15_64/df$tot
df$tot_65__ <- df$tot_65__/df$tot

df$male_0_14 <- df$male_0_14/df$tot_male 
df$male_15_64 <- df$male_15_64/df$tot_male 
df$male_65__ <- df$male_65__/df$tot_male 

df$fem_0_14 <- df$fem_0_14/df$tot_fem 
df$fem_15_64 <- df$fem_15_64/df$tot_fem 
df$fem_65__ <- df$fem_65__/df$tot_fem 

### OLS models

formula = inpost ~ dhl + dpd + fedex + ruch + ups +
            log(tot + 1) + fem_ratio + 
            buildings + shops + parks + forests + schools + 
            railways + cycleways + parkings + crossings + bus_stops

ols <- lm(formula, data=df)
summary(ols)
car::vif(ols)
bptest(ols) 
resettest(ols, power=2, type="regressor")
lm.morantest(ols, df.cont.listw) 

intercept_only <- lm(inpost~1, data=df)
all <- lm(formula, data=df)
both <- step(intercept_only, direction='both', scope=formula(all), trace=0)

Based on Stepwise Regression we obtained final formula of the model and we are able to run our final OLS with diagnostic part. We see that nearly all variables in the model are significant (of course due to the heteroscedasticity we use biased error variance estimator - we omitted application of the robust matrix due to lack of time). There is no collinearity in the model, but again we did not pass Ramsey and B-P tests. What is more residuals from the model are not spatially dependent. In case of interpretation we can see that: logistic companies have positive sign as expected; the greater the total density the more InPost pick-up points are stated; fem_ratio results is interesting because its sign is negative (more woman less pick-up points), in addition, the negative relationship between parks and pickup points is interesting but intuitive. There is positive relationship between pickup points and schools, parkings, cycleways and bus stops. It is also intuitive!

formula = formula(both)
ols <- lm(formula, data=df)
summary(ols)
cat("AIC", AIC(ols))
car::vif(ols)
bptest(ols) 
resettest(ols, power=2, type="regressor")
lm.morantest(ols, df.cont.listw) 

### Spatial models with 3, 2, or 1 spatial components

GNS <- sacsarlm(formula, data=df, listw=df.cont.listw, type="sacmixed", method="LU")
summary(GNS)

SAC <- sacsarlm(formula, data=df, listw=df.cont.listw)
summary(SAC)

SDM <- errorsarlm(formula, data=df, listw=df.cont.listw, etype="mixed")
summary(SDM)

SDEM <- errorsarlm(formula, data=df, listw=df.cont.listw, etype="emixed")
summary(SDEM)

SEM <- errorsarlm(formula, data=df, listw=df.cont.listw)
summary(SEM)

SAR <- lagsarlm(formula, data=df, listw=df.cont.listw)
summary(SAR)

SLX <- lmSLX(formula, data=df, listw=df.cont.listw)
summary(SLX)

Again SAR model seems to be the best in case of spatial terms significance and number of relevant variables! 
For Cracow BIC and AIC results also indicates SAR as the best model!

cat("ols",BIC(ols) ,'\n')
cat("GNS",BIC(GNS),'\n')
cat("SAC",BIC(SAC),'\n')
cat("SDM",BIC(SDM),'\n')
cat("SDEM",BIC(SDEM),'\n')
cat("SAR",BIC(SAR),'\n')
cat("SLX",BIC(SLX),'\n')
cat("SEM",BIC(SEM),'\n')
cat('\n')
cat("ols",AIC(ols) ,'\n')
cat("GNS",AIC(GNS),'\n')
cat("SAC",AIC(SAC),'\n')
cat("SDM",AIC(SDM),'\n')
cat("SDEM",AIC(SDEM),'\n')
cat("SAR",AIC(SAR),'\n')
cat("SLX",AIC(SLX),'\n')
cat("SEM",AIC(SEM),'\n')

Now we looked for better formula for SAR and we found it in the general to specific procedure.

formula_max = inpost ~ dhl + dpd + fedex + ruch + ups +
            log(tot + 1) + fem_ratio + 
            buildings + shops + parks + forests + schools + 
            railways + cycleways + parkings + crossings + bus_stops

formula_gen_to_specifc = inpost ~ dhl + dpd + ruch + ups +
            log(tot + 1) + fem_ratio + 
            parks + schools + 
            cycleways + parkings

SAR_g2s<- lagsarlm(formula_gen_to_specifc, data=df, listw=df.cont.listw)
summary(SAR_g2s) #the same as previously!

### Direct and indirect impacts

Now we move to the statistical inference part.

Let's start with logistic companies variables. We see that either in case of direct and indirect impacts all parameters for logistic companies (apart from Ruch total, and UPS indirect) are significant (at least around 10%). Internalization effects are greater than spill-over, but still spill-over effects are significant. In case of demographic variables we claim that total population density is important variable for the model in case of direct and indirect impact (has positive sign). Feminization ratio in case of direct impact is also significant and has negative sign (it is quite non intuitive). In case of point of interest data we see that number of schools, number of parks, number of cycleways and number of parkings are significant in the model with respectively positive, negative positive, and positive sign. We claim that it is intuitive.

# distribution of total impact 
# vector of traces of powers of a spatial weights matrix
# converting censored listw to CsparseMatrix form
W.c <- as(as_dgRMatrix_listw(df.cont.listw), "CsparseMatrix") 
# the default values for the number of powers is 30
trMat<-trW(W.c, type="mult") 

SAR_imp<-impacts(SAR_g2s, tr=trMat, R=200)
summary(SAR_imp, zstats=TRUE, short=TRUE)

## Conclusions from spatial dependence models

We combine our statistical inference for Warsaw and Cracow. Base of above research we can generally claim that InPost has deployed parcel pick-up points in line with the competitors - as the number of InPost pick-up points increases, the number of competitors' pick-up points increases (ignoring competitors not significant in the regression). We demonstrate that our control variables from the demographic group and points of interest are also significant, so InPost follows modeling logic and literature guidance for instance such variables are positively related with number of InPost's pick-up points: number of cycleways, number of parkings, number of schools, number of crossings. As it turned out a significant spatial effect in multivariate econometric models is Spatial Lag. It is quite intuitive that spillover effect exists for this problem.
